Save team when changed
After redux resolves dispatched actions, it calls subscribed functions.
Added a subscribe function for detecting team changes and ajaxing the
new team information to the server.

Switched ajax to axios. Pulled out csrf stuff.

Server side, when updating a record, had to convert its string id to an
IDObject in order for mongo to use it.